third_party/WebKit/LayoutTests/http/tests/feature-policy/fullscreen-allowed-by-container-policy-expected.txt

This is a testharness.js-based test.
FAIL Fullscreen is allowed on URL: resources/feature-policy-fullscreen.html with allowfullscreen = false assert_true: Document.fullscreenEnabled: expected true got false
FAIL Fullscreen is allowed on URL: http://localhost:8000/feature-policy/resources/feature-policy-fullscreen.html with allowfullscreen = false assert_true: Document.fullscreenEnabled: expected true got false
PASS Fullscreen is allowed on URL: resources/feature-policy-fullscreen.html with allowfullscreen = true 
PASS Fullscreen is allowed on URL: http://localhost:8000/feature-policy/resources/feature-policy-fullscreen.html with allowfullscreen = true 
Harness: the test ran to completion.
